[HUDI-7745] Move Hadoop-dependent util methods to hudi-hadoop-common
