+++ hudi-common/src/main/java/org/apache/hudi/common/fs/FSUtils.java
@@ -36,17 +36,13 @@ import org.apache.hudi.exception.HoodieValidationException;
 import org.apache.hudi.exception.InvalidHoodiePathException;
 import org.apache.hudi.metadata.HoodieTableMetadata;
 import org.apache.hudi.storage.HoodieStorage;
+import org.apache.hudi.storage.HoodieStorageUtils;
 import org.apache.hudi.storage.StorageConfiguration;
 import org.apache.hudi.storage.StoragePath;
 import org.apache.hudi.storage.StoragePathFilter;
 import org.apache.hudi.storage.StoragePathInfo;
 import org.apache.hudi.storage.inline.InLineFSUtils;
 
-import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.fs.FileStatus;
-import org.apache.hadoop.fs.FileSystem;
-import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.hdfs.DistributedFileSystem;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -71,8 +67,6 @@ import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import static org.apache.hudi.storage.HoodieStorageUtils.getStorageConfWithCopy;
-
 /**
  * Utility functions related to accessing the file storage.
  */
@@ -85,7 +79,6 @@ public class FSUtils {
   public static final Pattern LOG_FILE_PATTERN =
       Pattern.compile("^\\.(.+)_(.*)\\.(log|archive)\\.(\\d+)(_((\\d+)-(\\d+)-(\\d+))(.cdc)?)?");
   public static final Pattern PREFIX_BY_FILE_ID_PATTERN = Pattern.compile("^(.+)-(\\d+)");
-  private static final int MAX_ATTEMPTS_RECOVER_LEASE = 10;
 
   private static final String LOG_FILE_EXTENSION = ".log";
 
@@ -100,17 +93,6 @@ public class FSUtils {
    */
   public static boolean isTableExists(String path, HoodieStorage storage) throws IOException {
     return storage.exists(new StoragePath(path + "/" + HoodieTableMetaClient.METAFOLDER_NAME));
-  }
-
-  /**
-   * Makes path qualified w/ {@link FileSystem}'s URI
-   *
-   * @param fs   instance of {@link FileSystem} path belongs to
-   * @param path path to be qualified
-   * @return qualified path, prefixed w/ the URI of the target FS object provided
-   */
-  public static Path makeQualified(FileSystem fs, Path path) {
-    return path.makeQualified(fs.getUri(), fs.getWorkingDirectory());
   }
 
   /**
@@ -159,10 +141,6 @@ public class FSUtils {
     }
   }
 
-  public static long getFileSize(FileSystem fs, Path path) throws IOException {
-    return fs.getFileStatus(path).getLen();
-  }
-
   public static long getFileSize(HoodieStorage storage, StoragePath path) throws IOException {
     return storage.getPathInfo(path).getLength();
   }
@@ -187,13 +165,6 @@ public class FSUtils {
               path.getName()));
     }
     return datePartitions;
-  }
-
-  /**
-   * Given a base partition and a partition path, return relative path of partition path to the base path.
-   */
-  public static String getRelativePartitionPath(Path basePath, Path fullPartitionPath) {
-    return getRelativePartitionPath(new StoragePath(basePath.toUri()), new StoragePath(fullPartitionPath.toUri()));
   }
 
   public static String getRelativePartitionPath(StoragePath basePath, StoragePath fullPartitionPath) {
@@ -329,34 +300,12 @@ public class FSUtils {
     return matcher.group(3);
   }
 
-  /**
-   * Get the first part of the file name in the log file. That will be the fileId. Log file do not have instantTime in
-   * the file name.
-   */
-  public static String getFileIdFromLogPath(Path path) {
-    Matcher matcher = LOG_FILE_PATTERN.matcher(path.getName());
-    if (!matcher.find()) {
-      throw new InvalidHoodiePathException(path.toString(), "LogFile");
-    }
-    return matcher.group(1);
-  }
-
   public static String getFileIdFromLogPath(StoragePath path) {
     Matcher matcher = LOG_FILE_PATTERN.matcher(path.getName());
     if (!matcher.find()) {
       throw new InvalidHoodiePathException(path, "LogFile");
     }
     return matcher.group(1);
-  }
-
-  /**
-   * Check if the file is a base file of a log file. Then get the fileId appropriately.
-   */
-  public static String getFileIdFromFilePath(Path filePath) {
-    if (FSUtils.isLogFile(filePath)) {
-      return FSUtils.getFileIdFromLogPath(filePath);
-    }
-    return FSUtils.getFileId(filePath.getName());
   }
 
   public static String getFileIdFromFilePath(StoragePath filePath) {
@@ -447,11 +396,6 @@ public class FSUtils {
     return HoodieLogFile.LOG_FILE_PREFIX + suffix;
   }
 
-  public static boolean isBaseFile(Path path) {
-    String extension = getFileExtension(path.getName());
-    return HoodieFileFormat.BASE_FILE_EXTENSIONS.contains(extension);
-  }
-
   public static boolean isBaseFile(StoragePath path) {
     String extension = getFileExtension(path.getName());
     return HoodieFileFormat.BASE_FILE_EXTENSIONS.contains(extension);
@@ -463,10 +407,6 @@ public class FSUtils {
         ? InLineFSUtils.getOuterFilePathFromInlinePath(logPath).getName() : logPath.getName());
   }
 
-  public static boolean isLogFile(Path logPath) {
-    return isLogFile(new StoragePath(logPath.getName()));
-  }
-
   public static boolean isLogFile(String fileName) {
     if (fileName.contains(LOG_FILE_EXTENSION)) {
       Matcher matcher = LOG_FILE_PATTERN.matcher(fileName);
@@ -475,38 +415,8 @@ public class FSUtils {
     return false;
   }
 
-  /**
-   * Returns true if the given path is a Base file or a Log file.
-   */
-  public static boolean isDataFile(Path path) {
-    return isBaseFile(path) || isLogFile(path);
-  }
-
   public static boolean isDataFile(StoragePath path) {
     return isBaseFile(path) || isLogFile(path);
-  }
-
-  /**
-   * Get the names of all the base and log files in the given partition path.
-   */
-  public static FileStatus[] getAllDataFilesInPartition(FileSystem fs, Path partitionPath) throws IOException {
-    final Set<String> validFileExtensions = Arrays.stream(HoodieFileFormat.values())
-        .map(HoodieFileFormat::getFileExtension).collect(Collectors.toCollection(HashSet::new));
-    final String logFileExtension = HoodieFileFormat.HOODIE_LOG.getFileExtension();
-
-    try {
-      return Arrays.stream(fs.listStatus(partitionPath, path -> {
-        String extension = FSUtils.getFileExtension(path.getName());
-        return validFileExtensions.contains(extension) || path.getName().contains(logFileExtension);
-      })).filter(FileStatus::isFile).toArray(FileStatus[]::new);
-    } catch (IOException e) {
-      // return empty FileStatus if partition does not exist already
-      if (!fs.exists(partitionPath)) {
-        return new FileStatus[0];
-      } else {
-        throw e;
-      }
-    }
   }
 
   public static List<StoragePathInfo> getAllDataFilesInPartition(HoodieStorage storage,
@@ -568,29 +478,6 @@ public class FSUtils {
     return Option.empty();
   }
 
-  /**
-   * When a file was opened and the task died without closing the stream, another task executor cannot open because the
-   * existing lease will be active. We will try to recover the lease, from HDFS. If a data node went down, it takes
-   * about 10 minutes for the lease to be recovered. But if the client dies, this should be instant.
-   */
-  public static boolean recoverDFSFileLease(final DistributedFileSystem dfs, final Path p)
-      throws IOException, InterruptedException {
-    LOG.info("Recover lease on dfs file {}", p);
-    // initiate the recovery
-    boolean recovered = false;
-    for (int nbAttempt = 0; nbAttempt < MAX_ATTEMPTS_RECOVER_LEASE; nbAttempt++) {
-      LOG.info("Attempt {} to recover lease on dfs file {}", nbAttempt, p);
-      recovered = dfs.recoverLease(p);
-      if (recovered) {
-        break;
-      }
-      // Sleep for 1 second before trying again. Typically it takes about 2-3 seconds to recover
-      // under default settings
-      Thread.sleep(1000);
-    }
-    return recovered;
-  }
-
   public static void createPathIfNotExists(HoodieStorage storage, StoragePath partitionPath)
       throws IOException {
     if (!storage.exists(partitionPath)) {
@@ -600,10 +487,6 @@ public class FSUtils {
 
   public static Long getSizeInMB(long sizeInBytes) {
     return sizeInBytes / (1024 * 1024);
-  }
-
-  public static Path constructAbsolutePathInHadoopPath(String basePath, String relativePartitionPath) {
-    return new Path(constructAbsolutePath(basePath, relativePartitionPath).toUri());
   }
 
   public static StoragePath constructAbsolutePath(String basePath, String relativePartitionPath) {
@@ -640,13 +523,6 @@ public class FSUtils {
     int offset = StringUtils.isNullOrEmpty(partition)
         ? (filePathWithPartition.startsWith("/") ? 1 : 0) : partition.length() + 1;
     return filePathWithPartition.substring(offset);
-  }
-
-  /**
-   * Get DFS full partition path (e.g. hdfs://ip-address:8020:/<absolute path>)
-   */
-  public static String getDFSFullPartitionPath(FileSystem fs, Path fullPartitionPath) {
-    return fs.getUri() + fullPartitionPath.toUri().getRawPath();
   }
 
   /**
@@ -692,27 +568,6 @@ public class FSUtils {
       throw new HoodieIOException(ioe.getMessage(), ioe);
     }
     return false;
-  }
-
-  public static <T> Map<String, T> parallelizeFilesProcess(
-      HoodieEngineContext hoodieEngineContext,
-      FileSystem fs,
-      int parallelism,
-      SerializableFunction<Pair<String, StorageConfiguration<Configuration>>, T> pairFunction,
-      List<String> subPaths) {
-    Map<String, T> result = new HashMap<>();
-    if (subPaths.size() > 0) {
-      StorageConfiguration<Configuration> conf = getStorageConfWithCopy(fs.getConf());
-      int actualParallelism = Math.min(subPaths.size(), parallelism);
-
-      hoodieEngineContext.setJobStatus(FSUtils.class.getSimpleName(),
-          "Parallel listing paths " + String.join(",", subPaths));
-
-      result = hoodieEngineContext.mapToPair(subPaths,
-          subPath -> new ImmutablePair<>(subPath, pairFunction.apply(new ImmutablePair<>(subPath, conf))),
-          actualParallelism);
-    }
-    return result;
   }
 
   /**
@@ -775,59 +630,15 @@ public class FSUtils {
    */
   public static boolean deleteSubPath(String subPathStr, StorageConfiguration<?> conf, boolean recursive) {
     try {
-      Path subPath = new Path(subPathStr);
-      FileSystem fileSystem = subPath.getFileSystem(conf.unwrapAs(Configuration.class));
-      return fileSystem.delete(subPath, recursive);
+      StoragePath subPath = new StoragePath(subPathStr);
+      HoodieStorage storage = HoodieStorageUtils.getStorage(subPath, conf);
+      if (recursive) {
+        return storage.deleteDirectory(subPath);
+      }
+      return storage.deleteFile(subPath);
     } catch (IOException e) {
       throw new HoodieIOException(e.getMessage(), e);
     }
-  }
-
-  /**
-   * Lists file status at a certain level in the directory hierarchy.
-   * <p>
-   * E.g., given "/tmp/hoodie_table" as the rootPath, and 3 as the expected level,
-   * this method gives back the {@link FileStatus} of all files under
-   * "/tmp/hoodie_table/[*]/[*]/[*]/" folders.
-   *
-   * @param hoodieEngineContext {@link HoodieEngineContext} instance.
-   * @param fs                  {@link FileSystem} instance.
-   * @param rootPath            Root path for the file listing.
-   * @param expectLevel         Expected level of directory hierarchy for files to be added.
-   * @param parallelism         Parallelism for the file listing.
-   * @return A list of file status of files at the level.
-   */
-
-  public static List<FileStatus> getFileStatusAtLevel(
-      HoodieEngineContext hoodieEngineContext, FileSystem fs, Path rootPath,
-      int expectLevel, int parallelism) {
-    List<String> levelPaths = new ArrayList<>();
-    List<FileStatus> result = new ArrayList<>();
-    levelPaths.add(rootPath.toString());
-
-    for (int i = 0; i <= expectLevel; i++) {
-      result = FSUtils.parallelizeFilesProcess(hoodieEngineContext, fs, parallelism,
-          pairOfSubPathAndConf -> {
-            Path path = new Path(pairOfSubPathAndConf.getKey());
-            try {
-              FileSystem fileSystem = path.getFileSystem(pairOfSubPathAndConf.getValue().unwrap());
-              return Arrays.stream(fileSystem.listStatus(path))
-                .collect(Collectors.toList());
-            } catch (IOException e) {
-              throw new HoodieIOException("Failed to list " + path, e);
-            }
-          },
-          levelPaths)
-          .values().stream()
-          .flatMap(list -> list.stream()).collect(Collectors.toList());
-      if (i < expectLevel) {
-        levelPaths = result.stream()
-            .filter(FileStatus::isDirectory)
-            .map(fileStatus -> fileStatus.getPath().toString())
-            .collect(Collectors.toList());
-      }
-    }
-    return result;
   }
 
   public static List<StoragePathInfo> getAllDataPathInfo(HoodieStorage storage, StoragePath path)
@@ -843,35 +654,6 @@ public class FSUtils {
       }
     }
     return pathInfoList;
-  }
-
-  public static Map<String, Boolean> deleteFilesParallelize(
-      HoodieTableMetaClient metaClient,
-      List<String> paths,
-      HoodieEngineContext context,
-      int parallelism,
-      boolean ignoreFailed) {
-    return FSUtils.parallelizeFilesProcess(context,
-        (FileSystem) metaClient.getStorage().getFileSystem(),
-        parallelism,
-        pairOfSubPathAndConf -> {
-          Path file = new Path(pairOfSubPathAndConf.getKey());
-          try {
-            FileSystem fs = (FileSystem) metaClient.getStorage().getFileSystem();
-            if (fs.exists(file)) {
-              return fs.delete(file, false);
-            }
-            return true;
-          } catch (IOException e) {
-            if (!ignoreFailed) {
-              throw new HoodieIOException("Failed to delete : " + file, e);
-            } else {
-              LOG.warn("Ignore failed deleting : " + file);
-              return true;
-            }
-          }
-        },
-        paths);
   }
 
   /**
