+++ hudi-hadoop-mr/src/main/java/org/apache/hudi/hadoop/utils/HoodieRealtimeRecordReaderUtils.java
@@ -21,7 +21,7 @@ package org.apache.hudi.hadoop.utils;
 import org.apache.hudi.common.config.HoodieConfig;
 import org.apache.hudi.common.config.HoodieMemoryConfig;
 import org.apache.hudi.common.model.HoodieRecord;
-import org.apache.hudi.common.util.ConfigUtils;
+import org.apache.hudi.common.util.HadoopConfigUtils;
 import org.apache.hudi.common.util.collection.Pair;
 import org.apache.hudi.exception.HoodieException;
 import org.apache.hudi.hadoop.fs.HadoopFSUtils;
@@ -79,7 +79,7 @@ public class HoodieRealtimeRecordReaderUtils {
     // jobConf.getMemoryForMapTask() returns in MB
     return (long) Math
         .ceil(Double.parseDouble(
-            ConfigUtils.getRawValueWithAltKeys(
+            HadoopConfigUtils.getRawValueWithAltKeys(
                     jobConf, HoodieMemoryConfig.MAX_MEMORY_FRACTION_FOR_COMPACTION)
                 .orElse(HoodieMemoryConfig.DEFAULT_MR_COMPACTION_MEMORY_FRACTION))
             * jobConf.getMemoryForMapTask() * 1024 * 1024L);
